requirements.txt
torch>=2.1.0
torchvision>=0.16.0
open_clip_torch>=2.24.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.66.0
rich>=13.7.0
transformers>=4.41.0
accelerate>=0.31.0
